backend/polling/requirements.txt
google-cloud-firestore
google-cloud-pubsub
python-dotenv
flask
schema
pytz
